Reorgonize test suite
